requirements.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.2.1
psycopg2-binary==2.9.10
redis==5.2.1
cryptography==44.0.0
python-dotenv==1.0.1
gunicorn==23.0.0
django-cors-headers==4.6.0
